refactor:  RoomMembersProvider and create useRoomMembers hook for context management
